Make max retries configurable and increase potential jitter (#5)
This will allow users to back off to arbitrary lengths, if they have heavy load environments

Increasing potential jitter to ungroup jobs from being in the same second as much as is reasonable, so a 2s jitter increases the likelihood that they do not all continuously poll on the same second. More-so each round.